query
Say you have three tables WORK, USERS, MANAGERS
WORK
– work_id
– user_id
– how_much

USERS
– user_id
– team

MANAGERS
– manager_id
– team

If I am a manager, write a select statement to retrieve the work of all users who belong to my team. The mapping of user to team and team to manager are defined in the USERS and MANAGERS table.